Descripción clara de los cambios realizados

+++ Models/DTOs/Pedidos/PedidoDetalleDto.cs
@@ -6,6 +6,6 @@ public class PedidoDetalleDto
 {
     public int? IdArticulo { get; set; }   // futuro lookup
     public string? Articulo { get; set; }  // texto libre por ahora
-    public int Cantidad { get; set; }
+    public Decimal Cantidad { get; set; }
     public ArticuloLookupDto? ArticuloObj { get; set; }
 }
